namespace 변경
data, designsystem, domain, ui 모듈은 core 모듈의 서브 모듈이므로 core 패키지 추가
