+++ core/data/src/main/kotlin/com/nexters/bandalart/android/core/data/datasource/GuestLoginTokenDataSource.kt
@@ -1,10 +1,8 @@
-package com.nexters.bandalart.android.domain.repository
+package com.nexters.bandalart.android.core.data.datasource
 
 import kotlinx.coroutines.flow.Flow
 
-interface GuestLoginTokenRepository {
-
+interface GuestLoginTokenDataSource {
   suspend fun setGuestLoginToken(guestLoginToken: String)
-
   fun getGuestLoginToken(): Flow<Result<String>>
 }
