+++ core/data/src/main/kotlin/com/nexters/bandalart/android/core/data/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
-package com.nexters.bandalart.android.data.di
+package com.nexters.bandalart.android.core.data.di
 
-import com.nexters.bandalart.android.data.repository.GuestLoginTokenRepositoryImpl
-import com.nexters.bandalart.android.domain.repository.GuestLoginTokenRepository
+import com.nexters.bandalart.android.core.data.repository.GuestLoginTokenRepositoryImpl
+import com.nexters.bandalart.android.core.domain.repository.GuestLoginTokenRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
