+++ core/data/src/main/kotlin/com/nexters/bandalart/android/core/data/repository/GuestLoginTokenRepositoryImpl.kt
@@ -1,0 +1,17 @@
+package com.nexters.bandalart.android.core.data.repository
+
+import com.nexters.bandalart.android.core.data.datasource.GuestLoginTokenDataSource
+import com.nexters.bandalart.android.core.domain.repository.GuestLoginTokenRepository
+import javax.inject.Inject
+import kotlinx.coroutines.flow.Flow
+
+class GuestLoginTokenRepositoryImpl @Inject constructor(private val dataSource: GuestLoginTokenDataSource) :
+  GuestLoginTokenRepository {
+  override suspend fun setGuestLoginToken(guestLoginToken: String) {
+    dataSource.setGuestLoginToken(guestLoginToken)
+  }
+
+  override fun getGuestLoginToken(): Flow<Result<String>> {
+    return dataSource.getGuestLoginToken()
+  }
+}
