+++ core/domain/src/main/kotlin/com/nexters/bandalart/android/core/domain/usecase/GetGuestLoginTokenUseCase.kt
@@ -1,6 +1,6 @@
-package com.nexters.bandalart.android.domain.usecase
+package com.nexters.bandalart.android.core.domain.usecase
 
-import com.nexters.bandalart.android.domain.repository.GuestLoginTokenRepository
+import com.nexters.bandalart.android.core.domain.repository.GuestLoginTokenRepository
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
